audio-analysis/requirements.txt
pyannote.audio
torchaudio
pydub
whisper
numpy